requirements.txt
PyGithub>=2.1.1
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.1

